Move package to v1_0
