--- instrumentation/armeria-0.99/library/armeria-0.99-library.gradle
@@ -1,12 +1,0 @@
-ext {
-  minJavaVersionForTests = JavaVersion.VERSION_1_8
-  javaSubPackage = 'armeria.v0_99'
-}
-
-apply from: "$rootDir/gradle/instrumentation-library.gradle"
-
-dependencies {
-  compileOnly group: 'com.linecorp.armeria', name: 'armeria', version: '0.99.8'
-
-  testImplementation project(':instrumentation:armeria-0.99:testing')
-}
