+++ instrumentation/armeria-1.0/auto/armeria-1.0-auto.gradle
@@ -1,0 +1,24 @@
+ext {
+  minJavaVersionForTests = JavaVersion.VERSION_1_8
+}
+
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+// TODO(anuraaga): Move into instrumentation.gradle
+archivesBaseName = projectDir.parentFile.name
+
+muzzle {
+  pass {
+    group = "com.linecorp.armeria"
+    module = "armeria"
+    versions = "[0.99.8,)"
+  }
+}
+
+dependencies {
+  implementation project(path: ':instrumentation:armeria-1.0:library', configuration: 'shadow')
+
+  compileOnly group: 'com.linecorp.armeria', name: 'armeria', version: '0.99.8'
+
+  testImplementation project(':instrumentation:armeria-1.0:testing')
+}
