+++ instrumentation/armeria-1.0/library/armeria-1.0-library.gradle
@@ -1,0 +1,12 @@
+ext {
+  minJavaVersionForTests = JavaVersion.VERSION_1_8
+  javaSubPackage = 'armeria.v1_0'
+}
+
+apply from: "$rootDir/gradle/instrumentation-library.gradle"
+
+dependencies {
+  compileOnly group: 'com.linecorp.armeria', name: 'armeria', version: '0.99.8'
+
+  testImplementation project(':instrumentation:armeria-1.0:testing')
+}
