+++ settings.gradle
@@ -46,9 +46,9 @@ include ':instrumentation:annotations'
 include ':instrumentation:apache-httpasyncclient-4.0'
 include ':instrumentation:apache-httpclient:apache-httpclient-2.0'
 include ':instrumentation:apache-httpclient:apache-httpclient-4.0'
-include ':instrumentation:armeria-0.99:auto'
-include ':instrumentation:armeria-0.99:library'
-include ':instrumentation:armeria-0.99:testing'
+include ':instrumentation:armeria-1.0:auto'
+include ':instrumentation:armeria-1.0:library'
+include ':instrumentation:armeria-1.0:testing'
 include ':instrumentation:aws-sdk:aws-sdk-1.11'
 include ':instrumentation:aws-sdk:aws-sdk-2.2:auto'
 include ':instrumentation:aws-sdk:aws-sdk-2.2:library'
